refactor: implement uniqueLibraries method to optimize library declarations in GameInstanceManifest
